Now displaying landing page on server startup. files renamed to not seem like placeholders
