+++ controllers/controller.js
@@ -4,11 +4,11 @@ var express = require("express")
 var router = express.Router();
 
 // Import the model to use its database functions.
-var game = require("../models/exampleModel.js");
+var game = require("../models/model.js");
 // Create all our routes and set up logic within those routes where required.
 //Login Page
 router.get("/", function(request, response){
-    response.render("index")
+    response.render("index", {layout: 'landing'})
     console.log("Working Login")
 })
